alright, gonna take a break (risky!)

+++ src/map/block.cs
@@ -1,5 +1,5 @@
 public class block {
-    public ushort index;
+    public byte index;
 
     public static block air = new() { index = 0 };
     public static block def = new() { index = 1 };
